### 2018-08-06 #243 @MahShaaban
Submission: colocr
Conduct Co-localization Analysis of Fluorescence Microscopy Images

```
Package: colocr
Title: Conduct Co-localization Analysis of Fluorescence Microscopy Images
Version: 0.1.0
Authors@R: person("Mahmoud", "Ahmed",
    email = "mahmoud.s.fahmy@students.kasralainy.edu.eg",
    role = c("aut", "cre"))
Description: Automate the co-localization analysis of fluoresence microscopy
  images. Selecting regions of interest, extract pixel intensities from
  the image channels and calculate different co-localization statistics.
Suggests: testthat,
    covr,
    devtools
Imports: imager,
  shiny,
  scales
```

[e.g., "data extraction, because the package parses a scientific data file format"]
data extraction

Biologists (no advanced R required)

The [EBImage](https://www.bioconductor.org/packages/release/bioc/html/EBImage.html) and the [imager](https://cran.r-project.org/web/packages/imager/index.html) packages contain the algorithms for dealing with images. This package uses [imager](https://cran.r-project.org/web/packages/imager/index.html) but is focused only on the functionality needed for conducting co-localization of multichannel images. In addition, the package provide a shiny app to inteactively perform the same analysis.

A pre-submission, https://github.com/ropensci/onboarding/issues/241
