### 2018-08-08 #243 @maelle
:wave: @MahShaaban

> I am already using the shinytest in the app

 Ah sorry! But I don't see where in the tests? In particular, "shinytest" isn't written in DESCRIPTION like "testthat" is?
> Code auto-generation sounds like a great idea, do you recommend any resource to get started on that?
No, I'll research that and update this thread (and the guide, good point!)

I've now sent you an invite to the Slack.
