### 2018-08-08 #243 @MahShaaban
Thanks for the slack invite, @maelle

Regarding the shiny tests. The tests are in the app directory, `inst/colocr_app/tests/`. The reason is I wanted to test the app on Travis separately. If necessary, I will move the tests to run with the package testing.
