### 2018-08-08 #243 @MahShaaban
I am also new to `shinytest`.

I wanted to test the app separately cuz I intended to make it stand-alone and I have a prototype of it on shinyapps.io, [here](https://mahshaaban.shinyapps.io/colocr_app2/).

I have the app as a separate repo on github, [here](https://github.com/MahShaaban/colocr_app). This repo is added to the package repo as a submodule. The app repo contains separate `.travis.yml` and `tests/`. This way I can link the app as a regular repo on Travis and it runs the tests in with every commit.
