### 2018-08-08 #243 @MahShaaban
Alright. Will do.
I think I can add the build status badge from travis, if this what you mean.

Thanks
