### 2018-08-08 #243 @maelle
Do you want to look into code generation before your package undergoes review? I tagged you in a commit, see links to apps generating code [here](https://github.com/ropensci/dev_guide/pull/76#issuecomment-411426196).
