### 2018-08-09 9cac90f
fixed gp, added app badge to readme
