+++ man/roi_show.Rd
@@ -13,15 +13,15 @@ roi_show(img, px, labels, ind = c(1, 2))
 
 \item{labels}{An object of class \code{\link[imager]{cimg}}}
 
-\item{ind}{A \code{numeric} object of length two. For the channel indicies.}
+\item{ind}{A \code{numeric} object of length two. For the channel indexes.}
 }
 \description{
-Show/highlight the selected regions of interest on differen image channels
+Show/highlight the selected regions of interest on different image channels
 }
 \details{
 Calling this function on an image and a \code{px}, \code{\link[imager]{pixset}}
 of the same dimensions returns four different plots. The original image, a
-a low reslution representation of the \code{\link[imager]{pixset}} and the
+a low resolution representation of the \code{\link[imager]{pixset}} and the
 two channels indicated through \code{ind} highlighted. Additionally, when
 labels are provided through \code{labels} the regions are individually
 labeled.
