### 2018-08-09 2e92532
added shiny app badges
