### 2018-08-09 #243 @MahShaaban
I made the following changes
* I add a table to the **Table View** in the app that includes the input parameters and can be used to reproduce the same output from the app. I will keep looking into the code generation for future improvement.
* Fixed the issues raised by `goodpractice`
* Fixed the spelling errors from `devtools::spell_check`
* Added the app Travis and shinyapp.io badges to `README.md`
