### 2018-08-09 #243 @maelle
Cool, I'll now look for reviewers.

Could you please add this badge to the README of your package:
