+++ README.md
@@ -3,6 +3,7 @@
 [![Coverage Status](https://img.shields.io/codecov/c/github/MahShaaban/colocr/master.svg)](https://codecov.io/github/MahShaaban/colocr?branch=master)
 [![Build Status](https://travis-ci.org/MahShaaban/colocr_app.svg?branch=master)](https://travis-ci.org/MahShaaban/colocr_app)
 [![status](https://img.shields.io/badge/shinyapps.io-running-green.svg)](https://mahshaaban.shinyapps.io/colocr_app/) 
+[![](https://badges.ropensci.org/243_status.svg)](https://github.com/ropensci/onboarding/issues/243)
 
 # colocr
 
