### 2018-08-10 #243 @maelle
### Editor checks:

---

#### Editor comments

Already tackled :wink:

---

Reviewers: @haozhu233 @seaaan
Due date: 2018-09-03
