### 2018-08-10 #243 @maelle
Thanks @haozhu233 and @seaaan for accepting to review this package! Your reviews are due on 2018-09-03.

Here are links to [our reviewer guide](https://ropensci.github.io/dev_guide/onboarding-guide-for-reviewers.html) and [review template](https://ropensci.github.io/dev_guide/reviewtemplate.html) (as you're both experienced rOpenSci reviewers/authors comments on the guide are welcome in e.g. Slack but don't feel obliged to review both the package and the reviewing guide :wink:).
