+++ R/methods.R
@@ -12,7 +12,10 @@
 #' @param tolerance A \code{numeric} to be passed to \code{\link[imager]{label}}
 #' @param n A \code{numeric} of the number of regions of interest
 #'
-#' @return A \code{\link[imager]{pixset}} or a labeled image \code{\link[imager]{cimg}}
+#' @return A \code{\link[imager]{cimg}}. The original input \code{img} with an
+#' additional attribute \code{label}. \code{label} is a \code{vector} of
+#' \code{integer}s. The labels for the selected regions of interests starts
+#' from 1 and 0 is ignored.
 #'
 #' @details The function applies several \code{\link{imager}} morphological
 #' manipulations to select the regions of interest. These include
@@ -24,20 +27,19 @@
 #' is used to determine if two pixels belong to the same region.
 #'
 #' @examples
-#' # load libraries
+#' # load required libraries
 #' library(imager)
-#' library(colocr)
+#' library(magrittr)
 #'
 #' # load images
 #' fl <- system.file('extdata', 'Image0001_.jpg', package = 'colocr')
 #' img <- load.image(fl)
 #'
-#' # choose parameters
-#' px <- roi_select(img, threshold = 90)
+#' # choose ROI
+#' newimg <- roi_select(img, threshold = 90)
 #'
-#' # highlight chosen region of interest
-#' plot(img)
-#' highlight(px)
+#' # check the ROI labels
+#' unique(attr(newimg, 'label'))
 #'
 #' @importFrom imager is.cimg grayscale threshold as.pixset shrink grow fill clean
 #'
@@ -56,7 +58,7 @@ roi_select.default <- function(img, ...) {
 
 #' @export
 roi_select.cimg <- function(img, threshold, shrink = 5, grow = 5, fill = 5,
-                                  clean = 5, tolerance = .1, n) {
+                            clean = 5, tolerance = .1, n) {
 
   if(!is.numeric(threshold)) {
     stop('threshold should be a numeric between 0 and 100.')
@@ -86,50 +88,49 @@ roi_select.cimg <- function(img, threshold, shrink = 5, grow = 5, fill = 5,
   # add labels when n is provided
   if(!missing(n)) {
     labs.px <- .labels_add(px.m, tolerance = tolerance, n = n)
-    return(labs.px)
+    newimg <- img
+    attr(newimg, 'label') <- as.numeric(labs.px)
   } else {
-    return(px.m)
+    newimg <- img
+    attr(newimg, 'label') <- as.numeric(px.m)
   }
+  return(newimg)
 }
 
 #' Show the selected regions of interest
 #'
 #' Show/highlight the selected regions of interest on different image channels
 #'
-#' @param img An object of class \code{\link[imager]{cimg}}
-#' @param px An object of class \code{\link[imager]{pixset}}
-#' @param labels An object of class \code{\link[imager]{cimg}}
+#' @param img A \code{\link[imager]{cimg}} obeject such as the one returned
+#' from \code{\link{roi_select}}
 #' @param ind A \code{numeric} object of length two. For the channel indexes.
 #'
 #' @return NULL
 #'
-#' @details Calling this function on an image and a \code{px}, \code{\link[imager]{pixset}}
-#' of the same dimensions returns four different plots. The original image, a
-#' a low resolution representation of the \code{\link[imager]{pixset}} and the
-#' two channels indicated through \code{ind} highlighted. Additionally, when
-#' labels are provided through \code{labels} the regions are individually
-#' labeled.
[…]
