+++ R/helpers.R
@@ -1,3 +1,86 @@
+#' Get pixel intensities
+#'
+#' Get the pixel intensities of certain image channels
+#'
+#' @param img An object of class \code{\link[imager]{cimg}}
+#' @param ind A \code{numeric} of length two for channel indexes
+#'
+#' @return A \code{list} of three items. The first two items are the values of
+#' the pixel intensities of the channels indicated by \code{ind}. The third is
+#' the labels of the individual regions of interest.
+#'
+#' @examples
+#' # load required libraries
+#' library(imager)
+#' library(magrittr)
+#'
+#' # load images
+#' fl <- system.file('extdata', 'Image0001_.jpg', package = 'colocr')
+#' img <- load.image(fl)
+#'
+#' # choose parameters
+#' int <- roi_select(img, threshold = 90) %>%
+#'   .intensity_get()
+#'
+#' @importFrom imager channel is.cimg as.pixset is.pixset cimg
+#'
+#' @export
+.intensity_get <- function(img, ind = c(1,2)) {
+  UseMethod('.intensity_get')
+}
+
+#' @export
+.intensity_get.default <- function(img, ...) {
+  warning(paste("img is of class",
+                class(img),
+                ". img should be a cimg object."))
+}
+
+#' @export
+.intensity_get.cimg <- function(img, ind = c(1,2)) {
+
+  # get labels from img
+  # transform labels to cimg object
+  labels <- attr(img, 'label')
+  dims <- dim(grayscale(img))
+  a <- array(labels, dim = dims)
+
+  px <- cimg(a)
+
+  # get channel intensities
+  img1.g <- channel(img, ind = ind[1])
+  img2.g <- channel(img, ind = ind[2])
+
+  # use labels to subset images when provided
+  if(is.cimg(px)) {
+    # subset and change images to numeric
+    img1.num <- as.numeric(img1.g[as.pixset(px)])
+    img2.num <- as.numeric(img2.g[as.pixset(px)])
+
+    # calculate correlations
+    f <- as.data.frame(px)
+    f <- f[f$value != 0,]
+    f <- f$value
+
+    # make list
+    pixel_int <- list(channel1 = img1.num,
+                      channel2 = img2.num,
+                      labels = f)
+  } else if(is.pixset(px)) {
+    # subset and change images to numeric
+    img1.num <- as.numeric(img1.g[px])
+    img2.num <- as.numeric(img2.g[px])
+
+    # make list
+    pixel_int <- list(channel1 = img1.num,
+                      channel2 = img2.num,
+                      labels = rep(1, length(img1.num)))
+  }
+
+  # retrun corr
+  return(pixel_int)
+}
+
 #' Label regions of interest
 #'
 #' Add labels to regions of interest in an image
@@ -12,6 +95,7 @@
 #' @importFrom imager label as.cimg
 #' @importFrom stats reorder
 #'
+#' @export
 .labels_add <- function(px, tolerance = .1, n = 1) {
   px.labs <- label(px, tolerance = tolerance)
   value <- as.data.frame(px.labs)$value
@@ -49,6 +133,8 @@
 #' g <- rnorm(10)
 #'
 #' .pearson(r, g)
+#'
+#' @export
 .pearson <- function(r, g) {
   rx <- r - mean(r)
   gx <- g - mean(g)
@@ -75,6 +161,8 @@
 #' g <- rnorm(10)
 #'
 #' .manders(r, g)
+#'
+#' @export
 .manders <- function(r, g) {
   nom <- sum(r * g)
   den <- sqrt(sum(r**2) * sum(g**2))
