### 2018-08-31 #243 @haozhu233
3
---

I reviewed the `development` branch of this package as it has a stand-alone folder for that shiny app. I noticed that there are some recent changes in the master branch but they are mostly about badges. I run all the checks on OSX 10.13.

This package provides a straightforward way of conducting co-localization calculation, a specific type of image analysis for microscopy images. There are only 5 main functions in this package but they are quite useful and easy to use. The author also provided a GUI in a shiny app which makes it even more user friendly. The package vignette was also well written and easy to follow. Although I personally don't have any experience with analyzing microscopy images, I was still able to follow the tutorial and achieve what was supposed to be done. I can imagine this package can be quite useful for this particular use case.

Here I'm listing out some of the most significant issues I found:
- Right now the source codes for the shiny app sit in the `inst/colocr_app` folder. This approach works but if the app is considered as one of the core functionality of this package, I would recommend to put these codes in the R folder and recode it in a way that fits a R package
- There are some issues in the shiny app itself. I will explain in details below.
- The package vignette has numerous issues. Most of these issues have been mentioned by @seaaan and I just want to echo them as well. I couldn't do any better.
- @seaaan's comment on the lack of functionality to work on multiple images at once also applies to the shiny app. It would be nice to let user upload a zip file and then process the images one by one.

There are also a few minor issues that might be improved. I'm going to mention them below as well.

#### R codes (minor issues)
1. The default value for `n` in `roi_select` was defined in `.labels_add` which is not an exported function. It confuses RStudio and promotes a code diagnosis warning that "the value of `n` is not defined".

2. In `coloc_test`, it would be nice to standardize the `type` options with the terms used in the vignette. For example, `"pearsons"` can become `"PCC"` or `"pcc"`, or vice versa. At the same time, the option `"all"` should be `"both"`.

3. The purpose of using `S3` method here seems to be just sending out a warning message if `class(img) != "cimg"`. It seems a little unnecessary. This logic can simply be replaced by a `if` statement which can even be modularized to a small utility function.

4. It might be a good idea to split the contents of `methods.R` into multiple R files, like roi.R & intensity.R, based on their purposes.

5. The function name `run_app` is too general. Consider name it differently.

#### Shiny App
1. When there is no image input, there are some error message on the screen. To solve this problem, it's recommended to put `req(img1())` at the very beginning of every `output` object. For, example,

```
output$image_plot <- renderPlot({
    req(input$image1)
    par(mfrow=c(2,2), mar = rep(1, 4))
    roi_show(img = img1(),
                   px = px(),
                   labels = labs.px())
  })
```

2. There is no need to define `img1`, `labs.px`, `px` and `corr` in `reactiveValues`. You can just start with things like this.

```
values <- reactiveValues()
  # load images
img1 <- reactive({
    load.image(input$image1$datapath) # note that I removed the value assignment
})
```

3. For your question here (https://github.com/MahShaaban/colocr/blob/636435f57d034e02a9bc05cc7bd5d646e5dcb970/inst/colocr_app/app.R#L149), you can just say `values$df <- data.frame()` and that's enough.

For these three issues above, check out this modified version: https://gist.github.com/haozhu233/7a02bb4a03ccf4b5ef63acf40c53bbc7

4. As I said above, it would be nice if users can upload a zipped file with many pictures. You can add two button for going back and forward and a simple counter.

5. It would be nice to add a download button to those two tables on the "Tabular View" tab.

6. The correlation texts could be reformatted and become more readable.

#### Others
1. I also feel like that a pkgdown site might be helpful according to the package guideline.
