### 2018-09-01 #243 @MahShaaban
Thanks @haozhu233 for your review. I will work on the changes you suggested.
Regarding your last comment, I made some changes to use the pipeable approach, [here](https://github.com/MahShaaban/colocr/tree/pipeable). I haven't merged the changes to the development branch yet. So please, let me know if you have any comments on that as well.
