+++ inst/colocr_app/tests/mytest.R
@@ -1,0 +1,24 @@
+app <- ShinyDriver$new("../")
+app$snapshotInit("mytest")
+
+app$uploadFile(image1 = system.file('extdata', 'Image0001_.jpg', package = 'colocr'))
+app$setInputs(threshold = 90)
+app$setInputs(clean = 8)
+app$setInputs(fill = 8)
+app$setInputs(shrink = 6)
+app$setInputs(roi_num = 2)
+app$setInputs(name = "LC3")
+app$setInputs(add = "click")
+app$setInputs(add2 = "click")
+app$uploadFile(image1 = system.file('extdata', 'Image0003_.jpg', package = 'colocr'))
+app$setInputs(roi_num = 3)
+app$setInputs(fill = 5)
+app$setInputs(clean = 10)
+app$setInputs(shrink = 8)
+app$setInputs(fill = 10)
+app$setInputs(name = "LC3 ")
+app$setInputs(add = "click")
+app$setInputs(add2 = "click")
+#app$snapshotDownload("download_stats")
+#app$snapshotDownload("download_inputs")
+app$snapshot()
