+++ inst/colocr_app/app.R
@@ -1,0 +1,250 @@
+library(shiny)
+library(imager)
+library(colocr)
+
+# Define UI for application that draws a histogram
+ui <- navbarPage(title = 'colocr',
+                 tabPanel('Main',
+                          sidebarLayout(
+                            sidebarPanel(
+                              tags$h3('Input Panel'),
+                              tags$p('Get started by uploading the merge image. Then adjust the
+                                      parameters to fit the regions of interest. Finally, assign a
+                                      name to probe used in this image to be used in the output'),
+                              tags$hr(),
+                              fileInput('image1', 'Merge Image'),
+                              tags$hr(),
+                              sliderInput('threshold', 'Threshold', 1, 100, 50, 1),
+                              sliderInput('shrink', 'Shrink', 1, 10, 5, 1),
+                              sliderInput('grow', 'Grow', 1, 10, 5, 1),
+                              sliderInput('fill', 'Fill', 1, 10, 5, 1),
+                              sliderInput('clean', 'Clean', 1, 10, 5, 1),
+                              sliderInput('tolerance', 'Tolerance', 0, .99, .1, .1),
+                              numericInput('roi_num', 'ROI Num', 1, 1, 50, 1),
+                              tags$hr(),
+                              textInput('name', 'Probe Name')
+                              ),
+                            mainPanel(
+                              fluidRow(
+                                tags$h2('What are the different tabs for?'),
+                                tags$br(),
+                                tags$p('Each of the below tabs provide a view of your image, data
+                                       and analysis output. The different tabs are connected and are
+                                       updated automatically whenever the input panel is used.'),
+                                tags$li('Select ROI: Choose regions of interst by adjusting the input
+                                        parameters.'),
+                                tags$li('Pixel Intensities: Check the scatter and density distribution
+                                         of the pixel intensities from the two channels.'),
+                                tags$li('Tabular Output: View the different colocalization
+                                        co-efficients in tabular format.'),
+                                tags$li('Graph View: View the co-localization co-efficients in graphical
+                                        format.'),
+                                tags$br(),
+                                tags$br(),
+                                tabsetPanel(
+                                  tabPanel('Select ROI',
+                                           plotOutput("image_plot"),
+                                           textOutput('cor')
+                                  ),
+                                  tabPanel('Pixel Intensities', plotOutput('scatter')),
+                                  tabPanel('Tabular View',
+                                           tags$br(),
+                                           tags$h3('Co-localization stats table.'),
+                                           actionButton('add', 'Add'),
+                                           actionButton('remove', 'Remove'),
+                                           downloadButton('download_stats', 'Download Table'),
+                                           tableOutput('tab'),
+                                           tags$br(),
+                                           tags$h3('Input parameters'),
+                                           actionButton('add2', 'Add'),
+                                           actionButton('remove2', 'Remove'),
+                                           downloadButton('download_inputs', 'Download Table'),
+                                           tableOutput('tab2')
+                                  ),
[…]
