+++ tests/testthat.R
@@ -1,4 +1,5 @@
 library(testthat)
+library(shinytest)
 library(colocr)
 
 test_check("colocr")
