### 2018-09-02 f03e160
changes to pass travis
