+++ man/run_app.Rd
@@ -1,0 +1,11 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/methods.R
+\name{run_app}
+\alias{run_app}
+\title{Run the shiny App}
+\usage{
+run_app()
+}
+\description{
+Run the shiny App
+}
