+++ DESCRIPTION
@@ -14,6 +14,7 @@ Description: Automate the co-localization analysis of fluorescence microscopy
 Encoding: UTF-8
 LazyData: true
 Suggests: testthat,
+    shinytest,
     covr,
     knitr,
     rmarkdown,
