+++ NAMESPACE
@@ -18,6 +18,7 @@ export(roi_check)
 export(roi_select)
 export(roi_show)
 export(roi_test)
+export(run_app)
 importFrom(graphics,lines)
 importFrom(graphics,plot)
 importFrom(imager,as.cimg)
@@ -35,5 +36,6 @@ importFrom(imager,label)
 importFrom(imager,shrink)
 importFrom(imager,threshold)
 importFrom(scales,alpha)
+importFrom(shiny,runApp)
 importFrom(stats,density)
 importFrom(stats,reorder)
