+++ tests/testthat/test-app.R
@@ -1,6 +1,7 @@
 context('test app')
 
 test_that("app works", {
+  library(shinytest)
   app_dir <- system.file('colocr_app/', package = 'colocr')
   expect_pass(testApp(app_dir, compareImages = FALSE))
 })
