+++ R/methods.R
@@ -2,20 +2,29 @@
 #'
 #' Select regions of interest in an image using different morphological operations
 #'
-#' @param img An object of class \code{\link[imager]{cimg}}
+#' @param img An object of class \code{\link[imager]{cimg}} or a \code{list} of
+#' multiple \code{\link[imager]{cimg}} items
 #' @param threshold A \code{numeric} to be passed to
-#' \code{\link[imager]{threshold}}
+#' \code{\link[imager]{threshold}} or a \code{vector} of values fo each
+#' image in \code{img}
 #' @param shrink A \code{numeric} to be passed to \code{\link[imager]{shrink}}
+#' or a \code{vector} of values fo each image in \code{img}
 #' @param grow A \code{numeric} to be passed to \code{\link[imager]{grow}}
+#' or a \code{vector} of values fo each image in \code{img}
 #' @param fill A \code{numeric} to be passed to \code{\link[imager]{fill}}
+#' or a \code{vector} of values fo each image in \code{img}
 #' @param clean A \code{numeric} to be passed to \code{\link[imager]{clean}}
+#' or a \code{vector} of values fo each image in \code{img}
 #' @param tolerance A \code{numeric} to be passed to \code{\link[imager]{label}}
+#' or a \code{vector} of values fo each image in \code{img}
 #' @param n A \code{numeric} of the number of regions of interest
+#' or a \code{vector} of values fo each image in \code{img}
 #'
 #' @return A \code{\link[imager]{cimg}}. The original input \code{img} with an
 #' additional attribute \code{label}. \code{label} is a \code{vector} of
 #' \code{integer}s. The labels for the selected regions of interests starts
-#' from 1 and 0 is ignored.
+#' from 1 and 0 is ignored. When \code{img} is a list, a \code{list} is
+#' returned.
 #'
 #' @details The function applies several \code{\link{imager}} morphological
 #' manipulations to select the regions of interest. These include
@@ -45,7 +54,7 @@
 #'
 #' @export
 roi_select <- function(img, threshold, shrink = 5, grow = 5, fill = 5,
-                             clean = 5, tolerance = .1, n) {
+                             clean = 5, tolerance = .1, n = 1) {
   UseMethod('roi_select')
 }
 
@@ -53,12 +62,12 @@ roi_select <- function(img, threshold, shrink = 5, grow = 5, fill = 5,
 roi_select.default <- function(img, ...) {
   warning(paste("img is of class",
                 class(img),
-                ". img should be a cimg object."))
+                ". img should be a cimg or a list of cimg objects."))
 }
 
 #' @export
 roi_select.cimg <- function(img, threshold, shrink = 5, grow = 5, fill = 5,
-                            clean = 5, tolerance = .1, n) {
+                            clean = 5, tolerance = .1, n = 1) {
 
   if(!is.numeric(threshold)) {
     stop('threshold should be a numeric between 0 and 100.')
@@ -86,24 +95,70 @@ roi_select.cimg <- function(img, threshold, shrink = 5, grow = 5, fill = 5,
   px.m <- clean(px.m, clean)
 
   # add labels when n is provided
-  if(!missing(n)) {
-    labs.px <- .labels_add(px.m, tolerance = tolerance, n = n)
-    newimg <- img
-    attr(newimg, 'label') <- as.numeric(labs.px)
-  } else {
-    newimg <- img
-    attr(newimg, 'label') <- as.numeric(px.m)
-  }
-  return(newimg)
+  labs.px <- .labels_add(px.m, tolerance = tolerance, n = n)
+  attr(img, 'label') <- as.numeric(labs.px)
+
+  # return object
+  return(img)
 }
+
+#' @export
+roi_select.list <- function(img, threshold, shrink = 5, grow = 5, fill = 5,
[…]
