+++ man/roi_check.Rd
@@ -7,10 +7,11 @@
 roi_check(img, ind = c(1, 2))
 }
 \arguments{
-\item{img}{A \code{\link[imager]{cimg}} obeject such as the one returned
-from \code{\link{roi_select}}}
+\item{img}{A \code{\link[imager]{cimg}} obeject or a \code{list} of multiple
+images such as the one returned from \code{\link{roi_select}}}
 
-\item{ind}{A \code{numeric} object of length two. For the channel indexes.}
+\item{ind}{A \code{numeric} object of length two. For the channel indexes.
+or a \code{list} of similar vectors for each of \code{img} items.}
 }
 \description{
 Show the pixel intensities of certain image channels

[…]
