### 2018-09-03 #243 @MahShaaban
I'd like first to thank you guys for your time and the thoughtful comments.

I've made a quite a few changes based on the comments and the suggestions above. So I will start by mentioning these then I will reply to each of the reviews separately.

The changes I made for these reviews are in a new branch called `pipeable`

To install this branch from github direclty please use

```r
devtools::install_github('MahShaaban/colocr@pipeable')
```

Since the **master** still behind these changes and has the app directory as a submodule, please used the following to pull the repo and view it locally

```bash
$ git clone --recursive-submodules https://github.com/MahShaaban/colocr
```

## Main changes

The main changes are:
* Change the names of some of the functions; `intensity_show` to `roi_check`, `colocr_test` to `roi_test` and `run_app` to `colocr_app`.
* Make it possible for the functions to work with pipes `%>%`. So typical function call can be as following

```r
img <- imager::load.image('path to image file')
img %>%
    roi_select(threshold = 50) %>%
    roi_test()
```
So all function takes a `cimg` as an input. `roi_select` returns a `cimg` with an **attribute** `label`.
`roi_show` and `roi_check` return lists of 4 and 2 plots, respectively. Finally, `roi_test` returns a `data.frame`.

* I added an S3 method for each of the functions to take a `list` of `cimg` objects and returns a `list`
of their regular return value. This way, users can process multiple images at once.

* I updated the vignette to describe this new changes. I also corrected the issues that were raised in the reviews.

* I updated the shinyapp to work with the new pipeable functions and added a few changes to it including the ability to upload multiple files at once.

* I added more tests and moved the app tests to run with these of the package.

Here I reply to the different issues raised by both reviewers

## First review (@seaaan)

### Reviewer comments

* The shiny app is currently included as a submodule in the **master** branch. So you might need to use the `--recursive-submodules` when pulling the repo to work with it locally. Other branchs don't have this issue.
* I fixed the minor issues with the vignette
* Now it's possible to process multiple images at once

### Documentation
* I added a link to the **imager** page with installation instruction.
* The reason for using **development** is that the shinyapp is included in the **master** as a submodule no a regular directory, so `devtools::install_github('MahShaaban/colocr')` would fail.
* The issue with the vignette was probably because of the same submodule issue. I think pulling the repo with `--recursive-submodules` would solve this issue.
* I prefer to link to the vignette on from CRAN after the package is accepted.
* I fixed the minor issues in the vignette contents, functional documentation, and examples

### Functionality
* I used `req()` in each of the instance where the app shos and error on start. Now it shows nothing until the required input is provided.
* Regarding correlating more than one channel. I haven't seen this analysis done or required before. I can't come up with a case where such correlation is required. Also, using the `ind` input in `roi_test` can be a way to do that. One can calculate the correlation for say the first and the second channel then the second to the third.
* The functions can deal with `list` of images as described above. I think this can do for processing a collection of images at once.
* I used the suggested API so that the functions can work with pipes as described above.
* I will be looking at the speed issues the very next thing
* I included a way of reproducing the results from the app. Basically, one can export the inputs that were provided interactively and use them to regenerate the results. I added a section in the vignette to describe this process.

## Second review (@haozhu233)

### Reviewer comments
* As I've written in an earlier comment, I am trying to keep the app directory separated from the rest of the code and self-contained to be able to test and deploy the app later on.
* I think the changes described below would address the rest of the comments.

### R code
* I fixed the issue of the `n` default value in `roi_select`.
* `roi_test` no returns a `data.frame` instead of a `list`. The columns in the `data.frame` correspond to the different co-localization stats which are named the same as the `type` argument; `pcc` and/or `moc`.
* I think using **S3** is much cleaner. This is especially the case after adding methods for `list` to enable processing lists of images.

### Shiny app

* I fixed the issue of the errors thrown by the app on start using `req()` as suggested.
* I used the suggested code to define the reactive values
* The app can handle multiple files upload. However, I could only figure a way to upload the image files and not a zip file.
* I added `downloadButton`s to the **Tabular View**

***

I really appreciate all the comments and the suggestions from both reviewers. The definately helped me improving the code.
