### 2018-09-03 #243 @haozhu233
@MahShaaban Thanks for the responses. I just checked the branch and it looks a lot better. Here are some new issues I found.

- In the shiny app, I can only upload one image every time
- For zipped file entries, there is a `unzip` function in R that lets you unzip files and returns the names (you might want to create a temp dir and then do that).
- In the shiny app, after you upload an image, the "graph view" panel is broken with the changes as I can see a `ERROR:need finite 'xlim' values` on my machine.
- In the shiny app, maybe you can use a linebreak to replace the "and" in "Average Pearson's Correlation Coefficient: 0.85 and Average Manders Overlap Coefficient: 1"
- When threshold is set to 100, there is an error message `threshold should be a numeric >= 0 and < 100.`
- You might want to import & export `%>%` from `magritter` so this package works with pipe naturally.
- You imported `shinyBS` in DESCRIPTION but didn't use it in any `R` code other than the shiny app. Maybe you should put it in the suggest list. Same for `magritter`. If you don't have a plan to handle the `%>%` in NAMESPACE, you can throw it in the suggest list.
