+++ DESCRIPTION
@@ -19,11 +19,11 @@ Suggests: testthat,
   knitr,
   rmarkdown,
   devtools,
-  purrr
+  purrr,
+  shinyBS,
+  magrittr
 RoxygenNote: 6.0.1
 Imports: imager,
   shiny,
-  shinyBS,
-  scales,
-  magrittr
+  scales
 VignetteBuilder: knitr
