### 2018-09-04 #243 @MahShaaban
Thanks, @haozhu233

* I am pretty sure the app accepts more than one file. I made sure `multiple = TRUE` in `fileInput()`, [here](https://github.com/MahShaaban/colocr/blob/d25353ec347f9c8e5d34ee115acaa56d7e4fd771/inst/colocr_app/app.R#L16). I also have a test case with two images and it passes fine. If you tried that on the version of the app hosted on shinyapps.io it will not work since I haven't deployed the recent changes.
* I fixed this issue of **Graph View** by adding the plot in an `if` statement. So it only show the output `if(nrow(values$df) >= 1)`, [here](https://github.com/MahShaaban/colocr/blob/d25353ec347f9c8e5d34ee115acaa56d7e4fd771/inst/colocr_app/app.R#L274)
* I tried to format the text output but adding `\n` didn't work. I will keep looking at it.
* @seaaan noticed unexpected behavior when `roi_select(..., threshold = 100)`, so I restricted the input range to < 100.
* I added both `magritter` and `shinyBS` to `Suggests` in the DESCRIPTION.
