### 2018-09-04 #243 @haozhu233
@MahShaaban

- I was using RStudio 1.1 and the file uploader there doesn't support multiple files. After I start to use the preview version, this bug is gone.
- Great!
- How about a `<br>`?
- How about restrict the range in the shiny app to 0-99?
- Great!

Another tiny issue, can you reformat the UI part to make the codes look a little bit more structured?
