### 2018-09-04 #243 @seaaan
Nice changes! I have a few suggestions, but mostly it all looks good to me.

### Processing multiple images
I like the functionality of processing multiple images.

It would be nice to return a data frame of one row per image instead of a list of data frames when `roi_test` is called on a list of images

Maybe provide a `load_images()` function that takes a directory (or list of files?) and calls `lapply(., imager::load.image)`. You could also provide `load_image()` that just wraps `imager::load.image()` so the user doesn't have to run `library(imager)`

### Vignette comments
"The reset of the analysis" --> the RESULT of the analysis?

"Arguable, selecting" --> ArguablY

"only semi-automate" --> only semi-automateS

The link to ImageJ doesn't work

"shrin" --> shrinK (in the argument to roi_select)

The example checking that the app and package output are equal returns `FALSE`. Use `all.equal()` instead of `==`.

"regioins" --> regions

The first colocalization benchmark score image gives statistics relatively close to the expected. The second answer is very different! Comment in the vignette on why that is. It seems like ideally you should be able to reproduce the expected results?

ppc -> pcc (last code box)

### Other
With regard to the @development and "no submodules!" on the README, I suggest that you explain on the README what those mean (since I didn't know, maybe other users will also wonder).

`roi_show()` returns (via `imager::highlight`) `NULL` (eg as displayed in the vignette). If you return `invisible(NULL)` instead, `NULL` won't print out but you'll still get the same effect. You could also invisibly return the img to continue the piping.
