### 2018-09-05 a6b1709
added image_load and minor changes
