+++ inst/colocr_app/app.R
@@ -4,126 +4,178 @@ library(imager)
 library(colocr)
 
 # Define UI for application that draws a histogram
-ui <- navbarPage(title = 'colocr',
-                 tabPanel('Main',
-                          sidebarLayout(
-                            sidebarPanel(
-                              tags$h3('Input Panel'),
-                              tags$p('Get started by uploading the merge image. Then adjust the
-                                      parameters to fit the regions of interest. Finally, assign a
-                                      name to probe used in this image to be used in the output'),
-                              tags$hr(),
-                              fileInput('image1', 'Merge Image', multiple = TRUE),
-                              bsTooltip('image1',
-                                        'Upload one or more merge images.',
-                                        'right', options = list(container = "body")),
-                              tags$hr(),
-                              sliderInput('threshold', 'Threshold', 1, 100, 50, 1),
-                              bsTooltip('threshold',
-                                        'Choose a threshold for excluding the background.',
-                                        'right', options = list(container = "body")),
-                              sliderInput('shrink', 'Shrink', 1, 10, 5, 1),
-                              bsTooltip('shrink',
-                                        'Shrink the selected area by eroding the bounderies around it.',
-                                        'right', options = list(container = "body")),
-                              sliderInput('grow', 'Grow', 1, 10, 5, 1),
-                              bsTooltip('grow',
-                                        'Grow the selected area by dilating the bounderies around it.',
-                                        'right', options = list(container = "body")),
-                              sliderInput('fill', 'Fill', 1, 10, 5, 1),
-                              bsTooltip('fill',
-                                        'Remove holes in the selected area.',
-                                        'right', options = list(container = "body")),
-                              sliderInput('clean', 'Clean', 1, 10, 5, 1),
-                              bsTooltip('clean',
-                                        'Remove small isolated parts in the selected area.',
-                                        'right', options = list(container = "body")),
-                              sliderInput('tolerance', 'Tolerance', 0, .99, .1, .1),
-                              bsTooltip('tolerance',
-                                        'Set value to determine which two neighboring pixels are in same selected area.',
-                                        'right', options = list(container = "body")),
-                              numericInput('roi_num', 'ROI Num', 1, 1, 50, 1),
-                              bsTooltip('roi_num',
-                                        'Select the number of regions. Default is one.',
-                                        'right', options = list(container = "body")),
-                              tags$hr(),
-                              textInput('name', 'Probe Name'),
-                              bsTooltip('name',
-                                        'Enter the name of the probe used for staining. A name for all images or one for each separated by a comma.',
-                                        'right', options = list(container = "body"))
-                              ),
-                            mainPanel(
-                              fluidRow(
-                                tags$h2('What are the different tabs for?'),
-                                tags$br(),
-                                tags$p('Each of the below tabs provide a view of your image, data
-                                       and analysis output. The different tabs are connected and are
-                                       updated automatically whenever the input panel is used.'),
-                                tags$li('Select ROI: Choose regions of interst by adjusting the input
-                                        parameters.'),
-                                tags$li('Pixel Intensities: Check the scatter and density distribution
[…]
