### 2018-09-05 #243 @MahShaaban
Thanks, @haozhu233

* I tried using `<br>`, it didn't work as expected.
* I restricted the input to `threshold` in the shiny app to 0:99.
* I restructured and commented the code for the UI. Hopefully, it's more readable now.
