### 2018-09-05 #243 @MahShaaban
Hey, @seaaan
Here are some changes I made in response to your last comment.

## Processing multiple images

* The reason I choose `roi_select` to return a `list` is to keep have the returned `data.frame` units unchanged regardless of the number of input images. Returning one row for each image is not possible when there are multiple regions of interest in each image. Also, to have a straightforward way to map the output back to the input without having to subset a `data.frame`. I think the only other option, is to return a `data.frame` in case of multiple input images is to add an extra column identifier. Which also mean I have to assign these identifiers or ask for them as input from the user.
* I add a function `image_load` that takes a path to one or more images and returns a `cimg` or a list of `cimg` objects.

## Vignette Comments
* I fixed the typos you referred to.
* I used `all.equal` instead of `==`
* The second example of the benchmarking is a particularly tricky image to process. I didn't expect that the results from the package will match that from the original source. Nevertheless, I tried it and didn't want to remove it just because the results were not as expected. I am not sure why this is the case! But one reason may be is that in the image the areas that are meant to be selected are outside this structure in the middle so `colocr` would fail to produce reliable stats.

## Other
* Once I merge the change I made in the past few days in **master**, the submodule thing won't be an issue. I will include the source code of the app directly in the package repo.
* I used `invisible(NULL)` as a return statement in both `roi_show` and `roi_check`
