### 2018-09-06 #243 @seaaan
Nice work! Just a couple of things:

In the vignette, when you introduce analyzing a collection of images, it would be better to show how to use `image_load` with multiple file paths than to create the collection using `list(img, img2)` like it is now.

Also in the vignette, in the section where you're reproducing the analysis with the app and with code, the results aren't actually equivalent (i.e. `all.equal(...)`) returns a difference. That's confusing! Why aren't they equivalent?
