### 2018-09-07 e9bc218
change loaded pkgs and using image_load
