### 2018-09-07 5de5429
use image_load in collection section
