### 2018-09-07 #243 @MahShaaban
Thanks @seaaan

* I updated the vignette to use `image_load` in the section for analyzing a collection of images.
* I actually get the expected results when I knit the vignette locally.

I am not sure what might be the reason why you get different results. This is the case at least after rounding to the 2nd or 3rd digits. One thing I can think of is that the system files in your local version of the package might be not updated.
