### 2018-09-10 #243 @seaaan
Okay, well it must be something on my end then. As long as it says `TRUE` in the CRAN vignette it should be fine. I don't have any further comments. Nice work!
