### 2018-09-14 #243 @maelle
I get this error for tests. Any clue? Is there any dependency with a minimal version not indicated in DESCRIPTION yet?

```r
test-app.R:6: error: app works
Can't find variable: $
1: expect_pass(testApp(app_dir, compareImages = FALSE)) at C:\Users\Maelle\Documents\ropensci\colocr/tests/testthat/test-app.R:6
2: testApp(app_dir, compareImages = FALSE)
3: lapply(found_testnames, function(testname) {
       withr::local_dir(testsDir)
       withr::local_envvar(c(RSTUDIO = ""))
       withr::local_options(list(shinytest.app.dir = "appdir"))
       gc()
       env <- new.env(parent = .GlobalEnv)
       if (!quiet) {
           message(testname, " ", appendLF = FALSE)
       }
       source(testname, local = env)
   })
4: FUN(X[[i]], ...)
5: source(testname, local = env)
6: withVisible(eval(ei, envir))
7: eval(ei, envir)
8: eval(ei, envir)
9: ShinyDriver$new("../") at one_image.R:1
10: .subset2(public_bind_env, "initialize")(...)
11: sd_initialize(self, private, path, loadTimeout, checkNames, debug, phantomTimeout = phantomTimeout,
       seed = seed, cleanLogs = cleanLogs, shinyOptions = shinyOptions)
12: private$web$executeScript(js)
13: session_executeScript(self, private, script, ...)
14: private$makeRequest("EXECUTE SCRIPT", list(script = script, args = args))
15: session_makeRequest(self, private, endpoint, data, params, headers)
16: report_error(response)
√ |  6       | helpers.R [4.8 s]
√ | 32       | pipeable [205.1 s]

== Results =====================================================================
Duration: 223.1 s

OK:       38
Failed:   1
Warnings: 0
Skipped:  0
```
