### 2018-09-14 #243 @MahShaaban
I can't really understand what the problem is from the error.
I have merged the branch we've been woking on to the master and it passed the checks locally and on travis. You may try a fresh clone and see if the error goes away.
