### 2018-09-14 #243 @maelle
I get the same error with `shinytest` CRAN version.

``` r
library(shinytest)
#> Warning: package 'shinytest' was built under R version 3.5.1
#> Warning: package 'shinytest' was built under R version 3.5.1
  app_dir <- system.file('colocr_app/', package = 'colocr')
  expect_pass(testApp(app_dir, compareImages = FALSE))
#> Running one_image.R
#> Error in session_makeRequest(self, private, endpoint, data, params, headers): Can't find variable: $
```

Created on 2018-09-14 by the [reprex package](http://reprex.tidyverse.org) (v0.2.0).

<summary>Session info</summary>

``` r
devtools::session_info()
#> Session info -------------------------------------------------------------
#>  setting  value
#>  version  R version 3.5.0 (2018-04-23)
#>  system   x86_64, mingw32
#>  ui       RTerm
#>  language (EN)
#>  collate  English_United States.1252
#>  tz       Europe/Paris
#>  date     2018-09-14
#> Packages -----------------------------------------------------------------
#>  package    * version    date       source
#>  assertthat   0.2.0      2017-04-11 CRAN (R 3.5.0)
#>  backports    1.1.2      2017-12-13 CRAN (R 3.5.0)
#>  base       * 3.5.0      2018-04-23 local
#>  base64enc    0.1-3      2015-07-28 CRAN (R 3.5.0)
#>  callr        3.0.0      2018-08-24 CRAN (R 3.5.1)
#>  compiler     3.5.0      2018-04-23 local
#>  crayon       1.3.4      2017-09-16 CRAN (R 3.5.0)
#>  curl         3.2        2018-03-28 CRAN (R 3.5.0)
#>  datasets   * 3.5.0      2018-04-23 local
#>  debugme      1.1.0      2017-10-22 CRAN (R 3.5.0)
#>  devtools     1.13.5     2018-02-18 CRAN (R 3.5.0)
#>  digest       0.6.16     2018-08-22 CRAN (R 3.5.1)
#>  evaluate     0.11       2018-07-17 CRAN (R 3.5.1)
#>  graphics   * 3.5.0      2018-04-23 local
#>  grDevices  * 3.5.0      2018-04-23 local
#>  htmltools    0.3.6.9003 2018-09-04 Github (rstudio/htmltools@a9e969f)
#>  httpuv       1.4.4.1    2018-06-18 CRAN (R 3.5.0)
#>  httr         1.3.1      2017-08-20 CRAN (R 3.5.0)
#>  jsonlite     1.5        2017-06-01 CRAN (R 3.5.0)
#>  knitr        1.20       2018-02-20 CRAN (R 3.5.0)
#>  later        0.7.3      2018-06-08 CRAN (R 3.5.0)
#>  magrittr     1.5        2014-11-22 CRAN (R 3.5.0)
#>  memoise      1.1.0      2017-04-21 CRAN (R 3.5.0)
#>  methods    * 3.5.0      2018-04-23 local
#>  mime         0.5        2016-07-07 CRAN (R 3.5.0)
#>  parsedate    1.1.3      2017-03-02 CRAN (R 3.5.0)
#>  pingr        1.1.2      2017-03-02 CRAN (R 3.5.0)
#>  png          0.1-7      2013-12-03 CRAN (R 3.5.0)
#>  processx     3.2.0      2018-08-16 CRAN (R 3.5.1)
#>  promises     1.0.1      2018-04-13 CRAN (R 3.5.0)
#>  ps           1.1.0      2018-08-10 CRAN (R 3.5.1)
#>  R6           2.2.2      2017-06-17 CRAN (R 3.5.0)
#>  Rcpp         0.12.18    2018-07-23 CRAN (R 3.5.0)
#>  rematch      1.0.1      2016-04-21 CRAN (R 3.4.0)
#>  rlang        0.2.2      2018-08-16 CRAN (R 3.5.1)
#>  rmarkdown    1.10       2018-06-11 CRAN (R 3.5.0)
#>  rprojroot    1.3-2      2018-01-03 CRAN (R 3.4.3)
#>  shiny        1.1.0      2018-05-17 CRAN (R 3.5.0)
#>  shinytest  * 1.3.0      2018-05-07 CRAN (R 3.5.1)
#>  showimage    1.0.0      2018-01-24 CRAN (R 3.5.0)
#>  stats      * 3.5.0      2018-04-23 local
#>  stringi      1.2.4      2018-07-23 local
#>  stringr      1.3.1      2018-05-10 CRAN (R 3.5.0)
#>  testthat     2.0.0      2017-12-13 CRAN (R 3.5.0)
#>  tools        3.5.0      2018-04-23 local
#>  utils      * 3.5.0      2018-04-23 local
#>  webdriver    1.0.5      2018-04-11 CRAN (R 3.5.0)
#>  withr        2.1.2      2018-03-15 CRAN (R 3.4.4)
#>  xtable       1.8-2      2016-02-05 CRAN (R 3.4.0)
#>  yaml         2.2.0      2018-07-25 CRAN (R 3.5.1)
```
