### 2018-09-14 #243 @maelle
Oops so after re-installing `htmltools` (it was somehow corrupted), I was able to run the test. It told me the output was different. For the app screenshot it seems like a font difference. However it also shows me different results.

![image](https://user-images.githubusercontent.com/8360597/45571437-77bf4980-b866-11e8-8d84-f287368a608d.png)

@MahShaaban could you please add testing on Appveyor to see what you get there? Sorry to notice all of this only now.
