### 2018-09-14 #243 @maelle
I actually see you have a failing Appveyor build, but it seems due to a wrong Appveyor config, which prevents further tests from running. I'll stop exploring on my own system now, I hope you can get the tests passing on Appveyor.
