### 2018-09-14 #243 @maelle
Besides, the link in your shinyapps.io badge, https://mahshaaban.shinyapps.io/colocr_app/, has a 404 status.
