### 2018-09-14 #243 @maelle
Reg the Appveyor build this `webdriver` setup might help https://github.com/rstudio/webdriver/pull/29/files
