### 2018-09-15 e0c6e5e
fixed app link and add files for cran
