+++ vignettes/using_colocr.Rmd
@@ -31,8 +31,8 @@ A few R packages are available for conducting image analysis, which is a very wi
 The package development version is available at [github](https://github.com/MahShaaban/colocr).
 
 ```{r install_github, eval=FALSE}
-# install from github (pipeable branch)
-devtools::install_github('MahShaaban/colocr@pipeable')
+# install from github
+devtools::install_github('MahShaaban/colocr')
 ```
 
 This package depends on `imager` which has some external dependencies. The instructions for
