+++ NEWS.md
@@ -1,0 +1,6 @@
+# colocr 0.1.0
+
+* Added a `NEWS.md` file to track changes to the package.
+
+
+
