### 2018-09-15 #243 @MahShaaban
I fixed the link.
I can't get the tests to run on **appveyor**. I had a look at the issue from the `webdriver` repo. I don't get how they got it to pass, but if I have to guess I think they skipped that part of the test!
