### 2018-09-15 #243 @MahShaaban
I think it passed on **Travis** because I use `compareImages = FALSE` in `inst/colocr/runt_tests.R`. I use this option because these images are screenshots and would be different from an OS to another.
