### 2018-09-15 #243 @maelle
Yes but what about the average PCC etc? (Am still new to `shinytest`)
