### 2018-09-17 b541054
install phantomjs in appveyor
