### 2018-09-20 #243 @maelle
:wave: @MahShaaban!

My session info was in https://github.com/ropensci/onboarding/issues/243#issuecomment-421452708

Do you have access to a Windows machine?
