### 2018-09-20 #243 @MahShaaban
Thanks @maelle
I checked the package versions, they are comparable to the ones I have and still can't reproduce the error. Assuming that the tests fail only on windows, what would be the way forward?
I don't have an easy access to a windows machine, and I never really used R on one.
