### 2018-09-20 #243 @maelle
I've just had a look, can't do much more at the moment.

* Does the output in artifacts https://ci.appveyor.com/project/MahShaaban/colocr/build/artifacts help at all?

* Could you write lines of code that I could run to give you the output of, if that helps?

* I was hoping the tests of shinytest on Appveyor would help but I can't access them https://ci.appveyor.com/project/rstudio/shinytest - maybe you could request access?
