### 2018-09-20 #243 @maelle
I'm totally unexperimented with your app, hence my poor debugging. I was thinking that maybe when you set one of the input using the shinytest commands, the cursors move more or less than what you get on mac/Linux?

Btw on Travis do you test on on Linux? It'd be worth adding a mac build, just to see whether you only get the issue for Windows.
