### 2018-09-20 #243 @maelle
Or maybe it's due to a waiting time that's not enough on Windows? Could you try adding waiting time before the snapshot? Or even adding time between all commands?

End of my suggestions for today, sorry about that.
