### 2018-09-20 #243 @maelle
Maybe useful https://rstudio.github.io/shinytest/articles/in-depth.html#getting-input-output-and-export-values (to check the inputs have been set)
