### 2018-09-21 16ca5b2
increase test timeout in two_images apptest
