### 2018-09-21 #243 @MahShaaban
Hey @maelle
Here is what I tried recently.
* I already checked the artifacts and found nothing useful. Only the error message and the difference between the output and the expected.
* I am not sure how the tests of shiny test can be useful. I clicked the link and I couldn't access them either.
* On Travis I only test on Linux. I agree with your suggestion, maybe testing Mac can give a clue. However, I run the test on my mac and they pass with no changes!
* I increased the wait time for the steps that take long to update the output and this had no effects. I still get the same error and differences on appveyor. Check this [last commit](16ca5b2f68e34a1ef0ba98259eb876f5c2d2c2c4)
* I checked the link in your last comment and it has no mention of Linux/windows discrepancies that could be of help.

Here is an idea.
Since you can reproduce the errors on your local windows machine. I think it would be useful to see if the tests pass if you run them for the first time. To do that you need to remove the expected output folders from the app tests folder `inst/colocr/tests/*-expected/`. Once those two folders are removed, you can run the tests from the app directory `shinytest::testApp()`. This will give a message saying the tests are running for the first time and the expected images are being created. We can then compare the logs.
