### 2018-09-22 18e90c2
add test to reproduce the app output
