### 2018-09-22 c5e5d7d
add tests using imager only
