### 2018-09-22 #243 @MahShaaban
I wrote two tests to check whether the output of the app is reproduced using the same input parameters. Onc test used the `colocr` functions and another without. The tests passed locally and on Travis, while the two of them failed on Appveyor. In addition to the fact that, these are the only tests that actually check the numeric outputs (the co-localization stats), I think that neither the app nor the `colocr` functions are the root of the problem.

The tests are in a file called `tests/test-reproduce_app.R` in this [last commit](c5e5d7ddb2ae3922ea29f1bfbfb79cbc077b700f)

 PS. @seaaan noticed before that the part of the vignette where I check the reproduction of the app output from the same input returns `FALSE`. The code chunk is `check_equal` at line 313 of the vignette. Was this happening on a windows machine?
