### 2018-09-22 #243 @maelle
Interesting. Can you add a more minimal example using imager and data that's not in colocr so that I might run it and we can post in imager repo?
