### 2018-09-24 #243 @MahShaaban
I am not sure how to make a minimal example in this case. So far, I've been checking the final outputs of either `colocr` or `imager` and the tests passes locally and on travis but not appveyor. I am using multiple `imager` functions, and this difference could be due to any of them. So,
* I am not sure how to identify the particular function that produces the difference.
* `imager` builds on travis, only to an earlier commit, but it doesn't have appveyor integration.
* I made the the same test in `colocr/test-reproduce_app.R` to run independently from `colocr` if that helps. Basically, I replaced the `system.file()` calls to `download.file()` from GitHub first and reading them.

My current thought is to build `imager` on travis and appveyor and run this test. Meanwhile I am trying to figure out a way to identify the one or more functions that is causing the issue. I am not sure this is the smarts way to do it, but I think I can save all intermediary objects from the test run in an R object and compare them to a test run on windows/appveyor.
