### 2018-09-24 #243 @MahShaaban
Sorry, I forgot to link to the test script in the `imager` fork.
[Here](https://gist.github.com/MahShaaban/3f075e00237896b40e2bb09cc77f298b), is a gist of the script. I updated the second revision to remove `testthat` calls.
