### 2018-09-28 #243 @MahShaaban
I traced the difference between the calculated correlations on ubuntu and windows to very first step in loading the images. I think this(dahtah/imager#41) is related, although neither the maintainer nor the user followed up on the issue yet!

I found that the pixel values of the images in the `colocr` package are loaded differently on the two platforms. This was at least in part for `jpeg::readJPEG()` which is used in `imager::load.image()` and `colocr::image_load()`. I noticed the same when tried with other images. Here is one from the `jpeg` package itself.

On ubuntu
```
> version
               _
platform       x86_64-pc-linux-gnu
arch           x86_64
os             linux-gnu
system         x86_64, linux-gnu
status
major          3
minor          5.1
year           2018
month          07
day            02
svn rev        74947
language       R
version.string R version 3.5.1 (2018-07-02)
nickname       Feather Spray
> packageVersion('jpeg')
[1] ‘0.1.8’
> fl <- system.file('img', 'Rlogo.jpg', package = 'jpeg')
> img <- jpeg::readJPEG(fl)
> mean(img)
[1] 0.7046421
```

On windows
```
> version
               _
platform       x86_64-w64-mingw32
arch           x86_64
os             mingw32
system         x86_64, mingw32
status
major          3
minor          5.1
year           2018
month          07
day            02
svn rev        74947
language       R
version.string R version 3.5.1 (2018-07-02)
nickname       Feather Spray
> packageVersion('jpeg')
[1] ‘0.1.8’
> fl <- system.file('img', 'Rlogo.jpg', package = 'jpeg')
> img <- jpeg::readJPEG(fl)
> mean(img)
[1] 0.7047047
```
Notice the difference starting at the 4th decimal point. I am showing the `mean` here, but I visually inspected the values themselves and they look different. Could that be due to instability/inaccuracies at the very small decimal points?

I couldn't go beyond that as `jpeg::readJPEG` itself is a call to a source file, part of `CImg` C++ library as far as I can tell.

```
> jpeg::readJPEG
function (source, native = FALSE)
.Call("read_jpeg", if (is.raw(source)) source else path.expand(source),
    native, PACKAGE = "jpeg")
<bytecode: 0x2d11910>
<environment: namespace:jpeg>
```
