### 2018-09-28 ea69a66
replaced load.image with image_read
