+++ man/dot-manders.Rd
@@ -1,0 +1,29 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/helpers.R
+\name{.manders}
+\alias{.manders}
+\title{Calculate Marnders Overlap Coefficient}
+\usage{
+.manders(r, g)
+}
+\arguments{
+\item{r}{A \code{numeric} vector}
+
+\item{g}{A \code{numeric} vector}
+}
+\value{
+A \code{numeric} of length one.
+}
+\description{
+Calculates the manders overlap coefficient between two numeric vectors
+}
+\examples{
+set.seed(123)
+r <- rnorm(10)
+
+set.seed(1234)
+g <- rnorm(10)
+
+.manders(r, g)
+
+}
