+++ NAMESPACE
@@ -39,9 +39,10 @@ importFrom(imager,highlight)
 importFrom(imager,is.cimg)
 importFrom(imager,is.pixset)
 importFrom(imager,label)
-importFrom(imager,load.image)
+importFrom(imager,magick2cimg)
 importFrom(imager,shrink)
 importFrom(imager,threshold)
+importFrom(magick,image_read)
 importFrom(magrittr,"%>%")
 importFrom(scales,alpha)
 importFrom(shiny,runApp)
