+++ DESCRIPTION
@@ -21,8 +21,9 @@ Suggests: testthat,
   devtools,
   purrr,
   shinyBS
-RoxygenNote: 6.0.1
+RoxygenNote: 6.1.0
 Imports: imager,
+  magick,
   shiny,
   scales,
   magrittr
