+++ README.md
@@ -1,7 +1,7 @@
 [![Travis-CI Build Status](https://travis-ci.org/MahShaaban/colocr.svg?branch=pipeable)](https://travis-ci.org/MahShaaban/colocr)
-[![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/github/MahShaaban/colocr?branch=pipeable&svg=true)](https://ci.appveyor.com/project/MahShaaban/colocr)
+[![Build status](https://ci.appveyor.com/api/projects/status/m23tb10jxnq8b6xn?svg=true)](https://ci.appveyor.com/project/MahShaaban/colocr)
 [![Coverage Status](https://img.shields.io/codecov/c/github/MahShaaban/colocr/pipeable.svg)](https://codecov.io/github/MahShaaban/colocr?branch=pipeable)
-[![Build Status](https://travis-ci.org/MahShaaban/colocr_app.svg?branch=pipeable)](https://travis-ci.org/MahShaaban/colocr_app)
+[![Build Status](https://travis-ci.org/MahShaaban/colocr_app.svg?branch=master)](https://travis-ci.org/MahShaaban/colocr_app)
 [![status](https://img.shields.io/badge/shinyapps.io-running-green.svg)](https://mahshaaban.shinyapps.io/colocr_app2/) 
 [![](https://badges.ropensci.org/243_status.svg)](https://github.com/ropensci/onboarding/issues/243)
 
