### 2018-09-28 #243 @MahShaaban
I used `image_read` from the `magick` package instead of `load.image` from `imager` and this seems to solve the issue of ubuntu/windows differences. Here,  MahShaaban/colocr#3
