### 2018-09-28 #243 @maelle
Cool! In that case why not switch the whole package to magick? 😉
