### 2018-09-28 #243 @jeroen
I would second maelle's suggestion to try and switch to magick. It is a much more comprehensive and well tested imaging toolkit. Is there particular functionality in imager that you are missing from magick?
