### 2018-09-29 #243 @MahShaaban
Thanks, @maelle @jeroen for the suggestion. I certainly don't mind looking into that.

Although the package currently relies heavily on `imager`, I don't mind switching to `magick`. I went through the `magick` vignette and I think the classes and the basic image transformations that I'd need are already there. However, I don't see the equivalent/alternatives to the **Morphological Operations** in `imager`, namely `shrink()`, `grow()`, `fill()` and `clean()`. Or am I missing something in `magick` that could replace these?

Here is the relevant part from the `NAMESPACE`

```
importFrom(imager,clean)
importFrom(imager,fill)
importFrom(imager,grow)
importFrom(imager,shrink)
importFrom(imager,threshold)
```
