### 2018-09-29 #243 @jeroen
Thanks, see also this issue: https://github.com/ropensci/magick/issues/136

In the latest dev version of magick you can find the morphology methods with `morphology_types()`:

```r
> morphology_types()
 [1] "Undefined"         "Correlate"         "Convolve"          "Dilate"
 [5] "Erode"             "Close"             "Open"              "DilateIntensity"
 [9] "ErodeIntensity"    "CloseIntensity"    "OpenIntensity"     "DilateI"
[13] "ErodeI"            "CloseI"            "OpenI"             "Smooth"
[17] "EdgeOut"           "EdgeIn"            "Edge"              "TopHat"
[21] "BottomHat"         "Hmt"               "HitNMiss"          "HitAndMiss"
[25] "Thinning"          "Thicken"           "Distance"          "IterativeDistance"
[29] "Voronoi"
```

I think the main features you use are:

 - shrink: `magick::image_morphology(method = 'Erode', ...)`
 - grow: magick::image_morphology(method = 'Dilate', ...)

I'm not sure what exactly `imager::clean` does under the hood, but the [imagemagick morphology manual](http://www.imagemagick.org/Usage/morphology/) explains several morphology methods that can be used for cleaning. We also have a function `magick::image_despeckle()`.
