### 2018-10-02 59acd6e
fixed ci links and added ropensci footer
