### 2018-10-02 #243 @MahShaaban
Thank you, everyone.
I transferred the repo, fixed the ci links and will look into the other suggestions.

I'd like to acknowledge your contributions @maelle, @haozhu233, @seaaan, and @jeroen if you don't mind.
