### 2018-10-02 #243 @maelle
Awesome!

Don't acknowledge _my_ contributions, as mentioned [here](https://ropensci.github.io/dev_guide/building.html#authorship) "Please do not list editors as contributors. Your participation in and contribution to rOpenSci is thanks enough!" :wink: (we mean it!)
