### 2018-10-13 #243 @MahShaaban
Thanks, @stefaniebutland for this opportunity. I'd like to write a blog post about `colocr`. I will read the guides first and get back to you to discuss it.
