### 2018-10-23 #243 @stefaniebutland
@MahShaaban What do you think about setting a deadline to submit a draft post? I'm happy to answer any questions you might have.
