### 2018-10-24 #243 @stefaniebutland
If you're referring to a tech note (https://ropensci.org/technotes/), they don't require scheduling on a certain day of the week so please submit your draft when ready and I'll review it soon after.

> adapt parts of the vignette that explains the goal of the package and how it with examples.

Sounds good. Make sure it's different enough from the vignette. Good if you can lay out one cool example of what you can do with the package, rather than giving several examples.
