### 2018-10-25 #243 @stefaniebutland
@MahShaaban, @maelle just reminded me that this is your *second* package onboarding! I'm quite curious about package authors' motivations to submit multiple packages e.g. are there diminishing returns on author's effort on subsequent submissions?

I know you indicated you prefer to write a tech note about `colocr`, but if it interests you and you see value in it for yourself, I'd love to read a blog post that features `colocr` as you described, but also reflects on your experiences and motivation for onboarding multiple packages.

Zero obligation to do more than you suggested! 😄
