### 2018-11-02 #243 @MahShaaban
Hey @stefaniebutland, I just submitted a PR with a first draft of the blog post, ropensci/roweb2#329. Please, let me know what you think.
